isolate client list component
